Hue Lights Commit: attempt 2: changed files now showing
Have finished creating the lights.js file for my website. Have added script functionality to the game in the main.js file so it works in conjunction to the Hue lights. When ever it is player 1's turn on the left pane, lights 1-3 will turn blue. Whenever it is player 2's turn on the right pane, lights 4-6 will turn blue.